app/auth/jwt_handler.py
from datetime import datetime, timedelta
from jose import jwt, JWTError
from app.config import JWT_SECRET, JWT_ALGORITHM

def create_access_token(user_id: int):
    payload = {
        "sub": str(user_id),
        "exp": datetime.utcnow() + timedelta(hours=1)
    }
    return jwt.encode(payload, JWT_SECRET, algorithm=JWT_ALGORITHM)

def verify_access_token(token: str):
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        return int(payload.get("sub"))

    except JWTError:
        return None