requirements.txt
fastapi==0.115.0
uvicorn==0.31.1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

python-dotenv==1.1.1

python-jose[cryptography]==3.3.0
cryptography==42.0.8

passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic==2.12.3
pydantic_core==2.41.4
